requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7

# FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Optional: OpenAI (for future AI integration)
# openai==1.3.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
